[1,2,3 단계]Web MVC  (#15)
* docs(): 1단계 요구사항 정리

* feat(): [1단계] 로그인 및 내 정보 조회하기 기능 구현

* feat(): [1단계] AuthService 구현으로 인한 리팩터링

* docs(): [2단계] 요구사항 정리

* feat(): [2단계] 비로그인 사용자는 예약이 불가능하도록 구현

* feat(): [2단계] 자신의 예약이 아닌 경우 예약 취소가 불가능하도록 구현

* docs(): [3단계] 요구사항 정리

* feat(): [3단계] 일반 멤버와 관리자 멤버를 구분

* feat(): [3단계] 관리자 추가 및 어드민 인수 테스트 작성

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -1,5 +1,8 @@
 package nextstep.reservation;
 
+import nextstep.auth.AuthenticationException;
+import nextstep.member.Member;
+import nextstep.member.MemberDao;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
 import nextstep.support.DuplicateEntityException;
@@ -14,14 +17,17 @@ public class ReservationService {
     public final ReservationDao reservationDao;
     public final ThemeDao themeDao;
     public final ScheduleDao scheduleDao;
+    public final MemberDao memberDao;
 
-    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, ScheduleDao scheduleDao) {
+    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, ScheduleDao scheduleDao, MemberDao memberDao) {
         this.reservationDao = reservationDao;
         this.themeDao = themeDao;
         this.scheduleDao = scheduleDao;
+        this.memberDao = memberDao;
     }
 
-    public Long create(ReservationRequest reservationRequest) {
+    public Long create(ReservationRequest reservationRequest, Long memberId) {
+        Member member = memberDao.findById(memberId);
         Schedule schedule = scheduleDao.findById(reservationRequest.getScheduleId());
         if (schedule == null) {
             throw new NullPointerException();
@@ -34,7 +40,7 @@ public class ReservationService {
 
         Reservation newReservation = new Reservation(
                 schedule,
-                reservationRequest.getName()
+                member
         );
 
         return reservationDao.save(newReservation);
@@ -49,12 +55,21 @@ public class ReservationService {
         return reservationDao.findAllByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, Long memberId) {
+        Member member = memberDao.findById(memberId);
         Reservation reservation = reservationDao.findById(id);
         if (reservation == null) {
             throw new NullPointerException();
         }
+        validateOwner(member, reservation);
 
         reservationDao.deleteById(id);
     }
+
+    private void validateOwner(Member member, Reservation reservation) {
+        if (!reservation.isOwner(member)) {
+            throw new AuthenticationException("자신의 예약이 아닌 경우 삭제할 수 없습니다.");
+        }
+    }
+
 }
